Restore the hardcoded JWT
After the restructuring of the JWT code, the auto-refresh stopped working.
Probably the long-term solution is to move to the new version of the google
sign in for iOS, but I need to see whether they will work with a webview so
that we don't get bumped from the play store again.

So let's restore the old, hardcoded JWT and see if this gets at least the
background sync to start working again.