Make the javascript calls to get private data also go through native code
The original idea was that the javascript implementation of the code would just
use the javascript HTTP implementations, and make a call to the native code to
get the JWT. As we can saw, this does not refresh the token correctly if it is
expired, which is why we went back to the old hardcoded method.

But then if we want to use the old hardcoded method from javascript, we need to
make calls back to native code.

This change implements plugin interfaces on both ios and android that make the
appropriate calls to the native code, and change the javascript to call it
appropriately.

This has not yet been tested, but let's get an initial version checked in.