Fix the issue with the authentication not working on startup
Basically, if the currAuth is NULL, then we read it from the keychain. But with
the previous structure, after reading it, we just exit since everything else
was in a giant else.

Now, once we read the currAuth, we use it in the rest of the code.
Wow. Simple restructuring. Much working.