Ensure that all data is sent over as UTF-8
On android, unless we explicitly serialize the data as UTF-8, it is serialized as ISO-8851.
On iOS, it is serialized as UTF-8 by default.

In both, it is probably best to explicitly mark the content as UTF-8 in the HTTP header

Bump up the version number to match

This fixes https://github.com/e-mission/e-mission-docs/issues/333

Testing done:
- on android: set the locale to French, set the date to Fevr and synced data
    - before this change, it did not work
    - after this change, it works
- on iOS, set the locale to French, set the date to Fevr and synced data
    - the formatted data did not have any words so even before this fix, it worked
    - this did not cause a regression