Create nested try blocks
Before this change, the resultString was defined inside the try block but we attempted to print it out in the catch block. 

This resulted in a compile error

```
> Task :app:compileDebugJavaWithJavac FAILED
/Users/kshankar/Desktop/data/e-mission/native_code_upgrade/platforms/android/app/src/main/java/edu/berkeley/eecs/emission/cordova/comm/CommunicationHelperPlugin.java:31: error: cannot find symbol
                              + "Response was not JSON: " + resultString
                                                            ^
  symbol: variable resultString
1 error
```

Fixing it by creating nested try/catch blocks where the outer blocks checks for generic exceptions, and the inner block only includes the JSON parsing statement, and checks for a JSON exception. This makes the raw (unparsed) result string available to the error message.